samples/multi-agent-travel-planner/requirements.txt
# Core LangChain packages
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# Azure packages
azure-monitor-opentelemetry>=1.0.0
langchain-azure-ai>=0.1.0

# OpenTelemetry packages
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Utility packages
python-dotenv>=1.0.0

# Web UI
streamlit>=1.28.0